requirements.txt
Flask==2.3.3
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
webcolors==1.13